+++ planifun/src/main/java/univ/rouen/planifun/app/editor/model/task/Task.java
@@ -14,7 +14,7 @@ public interface Task {
 
     public int getCompletionDate();
 
-    public int getProgressStatus();
+    public Double getProgressStatus();
 
     // COMMANDS
